backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
pydantic>=2.11.0
pydantic-settings>=2.10.0
email-validator==2.2.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
# Updated to resolve dependency conflicts
httpx>=0.24.0,<0.25.0  # Compatible with both openai and supabase
openai==1.3.7
langchain==0.0.350
qdrant-client==1.7.0
supabase==2.0.2
sentry-sdk==1.38.0
Pillow>=10.0.0,<11.0.0
aiofiles>=23.1.0,<24.0.0
